consolidate node services
